:recycle: refact(calculator) : 분리 - Calculator 클래스
